move hover_pie_test.js content to pie_test.js
- no need for two test suite for pie traces.
